Broken build and incremental compitation fixes.

+++ flexmojos-maven-plugin/src/main/java/net/flexmojos/oss/plugin/compiler/attributes/MavenFontsConfiguration.java
@@ -214,7 +214,8 @@ public class MavenFontsConfiguration
         File fontsSer = new File( outputDirectory, "fonts.ser" );
         try
         {
-            FileUtils.copyURLToFile( url, fontsSer );
+            if (!fontsSer.exists())
+                FileUtils.copyURLToFile( url, fontsSer );
         }
         catch ( IOException e )
         {
